Avoid allocation of unnecessary skiplist value.
Adding a new node to the cache has two phases: first create a node in
the "initializing" state and then later go back and finish
initialization. The "initializing" state is indicated by a meta bit
with the value of 0x1. That bit is cleared when initialization is
finished.

The skiplist SetMeta method has an optimization where if the new meta
value is > the existing value, no arena allocation is necessary. This
optimization could not be used when adding a new node, because the
initializing bit went from 0x1 to 0x0 (i.e. decreased).

The fix is to invert the meta bit so that it indicates a node in the
"initialized" state rather than in the "initializing" state. Now the
bit starts out as 0x0 and later gets updated to 0x1, which allows for
the SetMeta optimization.